chore: Changed all dates from Job to double type + freq in Server
Changed all date types and Units of Work for Job class from int to
double in order to avoid casting to integers and also to prepare for
exercise 3 which may now include real number for dates, units of work and
other parameters.

+++ codes/src/components/Job.java
@@ -2,14 +2,14 @@ package components;
 
 public class Job {
     private int id;
-    private int arrivalDate;
-    private int unitsOfWork;
-    private int deadline;
-    private int period;
+    private double arrivalDate;
+    private double unitsOfWork;
+    private double deadline;
+    private double period;
 
     /* ============== CONSTRUCTORS ============== */
-    public Job(int id, int arrivalDate, int unitsOfWork,
-               int deadline, int period){
+    public Job(int id, double arrivalDate, double unitsOfWork,
+               double deadline, double period){
         this.id = id;
         this.arrivalDate = arrivalDate;
         this.unitsOfWork = unitsOfWork;
@@ -27,11 +27,11 @@ public class Job {
 
     /* ============== GETTERS ============== */
     public int getId() { return id; }
-    public int getADeadline(){ return arrivalDate + deadline; }
-    public int getArrivalDate() { return arrivalDate; }
-    public int getPeriod() { return period; }
-    public int getRDeadline() { return deadline; }
-    public int getUnitsOfWork() { return unitsOfWork; }
+    public double getADeadline(){ return arrivalDate + deadline; }
+    public double getArrivalDate() { return arrivalDate; }
+    public double getPeriod() { return period; }
+    public double getRDeadline() { return deadline; }
+    public double getUnitsOfWork() { return unitsOfWork; }
 
     /* ============== SETTERS ============== */
     public void decrement(int consumedUnitsOfWork){
